downloadable cost report in Excel

+++ projects/reports.py
@@ -6,6 +6,7 @@ from datetime import datetime
 import math
 from django.conf import settings
 import json
+import pandas as pd
 
 from django.core.mail import send_mail
 from django.template.loader import render_to_string
@@ -14,15 +15,13 @@ logger = logging.getLogger(__name__)
 
 GB = 1024 * 1024 * 1024
 today = datetime.now()
-end_month = today.month
-start_year = today.year
-end_year = today.year
+
 
 def _researchstats_month(research_folder, year, month):
     return ResearchStats.objects.filter(research_folder=research_folder, collected__year=year,
                                         collected__month=month).order_by('collected').last()
 
-def _monthly_research_stats(folder, stats):
+def _monthly_research_stats(folder, stats, start_year, end_year, end_month):
     for year in range(start_year, end_year + 1):
         if year not in stats:
             stats[year] = {}
@@ -37,19 +36,19 @@ def _monthly_research_stats(folder, stats):
                     stats[year][month]['size'] += s.size + s.revision_size
     return stats
 
-def _yearly_research_stats(folder, stats):
+def _yearly_research_stats(folder, stats, start_year, end_year):
     for year in range(start_year, end_year + 1):
         if year not in stats:
             stats[year] = {}
-            s = ResearchStats.objects.filter(research_folder=folder, collected__year=year).order_by('collected').last()
-            if s is not None:
-                if 'size' not in stats[year]:
-                    stats[year]['size'] = s.size + s.revision_size
-                else:
-                    stats[year]['size'] += s.size + s.revision_size
+        s = ResearchStats.objects.filter(research_folder=folder, collected__year=year).order_by('collected').last()
+        if s is not None:
+            if 'size' not in stats[year]:
+                stats[year]['size'] = s.size + s.revision_size
+            else:
+                stats[year]['size'] += s.size + s.revision_size
     return stats
 
-def _vault_datasets_by_month(vault_folder, data):
+def _vault_datasets_by_month(vault_folder, data, start_year, end_year, end_month):
     for year in range(start_year, end_year + 1):
         if year not in data:
             data[year] = {}
@@ -106,7 +105,7 @@ def send_monthly_owner_reports():
             fp.write(msg_html)
         try:
             send_mail(
-                f'{datetime.datetime.now().strftime("%B")} Yoda usage report for "{project.title}"',
+                f'{datetime.datetime.now()} Yoda usage report for "{project.title}"',
                 msg_plain,
                 settings.REPORTS_MAIL_FROM,
                 [settings.REPORTS_MAIL_TEST_TO],
@@ -116,9 +115,7 @@ def send_monthly_owner_reports():
         except Exception as e:
             logger.error(f'unable to send monthly report email to {settings.REPORTS_MAIL_TEST_TO}, error: {e}')
 
-#send_monthly_owner_reports()
-
-def get_billing_data():
+def get_billing_data(start_year, end_year, end_month):
     projects = Project.objects.filter(delete_date__isnull=True).all()
     billing_data = {}
     for project in projects:
@@ -128,11 +125,15 @@ def get_billing_data():
         datasets = {}
         research_yearly = {}
         for f in rf:
-            research = _monthly_research_stats(f, research)
-            datasets = _vault_datasets_by_month(VaultFolder.objects.get(research_folder=f), datasets)
-            research_yearly = _yearly_research_stats(f, research_yearly)
+            print(f'{project.title} - {f.yoda_name}')
+            research = _monthly_research_stats(f, research, start_year, end_year, end_month)
+            datasets = _vault_datasets_by_month(VaultFolder.objects.get(research_folder=f), datasets, start_year, end_year, end_month)
+            research_yearly = _yearly_research_stats(f, research_yearly, start_year, end_year)
         billing_data[project.id] = {
-            'project': project.title,    
+            'project': project.title,
+            'usage_details': f'https://adminyoda.labs.vu.nl/projects/{project.id}',
+            'created': project.created.strftime("%Y-%m-%d"),
+            'deleted': project.delete_date.strftime("%Y-%m-%d") if project.delete_date else '',
             'owner_name': project.owner.firstname + ' ' + project.owner.lastname,
             'owner_vunetid': project.owner.vunetid,
             'budget_code': project.budget.code,
@@ -155,11 +156,10 @@ def calculate_research_cost(size, free_block = 500, first_block = 2048, first_bl
     
 
 def calculate_monthly_costs_per_project(year):
-    billing_data = get_billing_data()
+    billing_data = get_billing_data(start_year=year, end_year=year, end_month=12)
     for project in billing_data:
         total_research_cost = 0
         total_dataset_cost = 0
-        print(project)
         if year in billing_data[project]['research']:
             for month in billing_data[project]['research'][year]:
                 cost = 0
@@ -180,19 +180,25 @@ def calculate_monthly_costs_per_project(year):
                 billing_data[project]['yearly_research_cost'] = calculate_cost(billing_data[project]['research_yearly'][year]['size'])
         billing_data[project]['total_dataset_cost'] = total_dataset_cost
         billing_data[project]['total_research_cost'] = total_research_cost
-        billing_data[project]['total_cost'] = total_research_cost + total_dataset_cost
+        billing_data[project]['total_cost'] = round(total_research_cost + total_dataset_cost)
     return billing_data
 
 def yearly_billing_report(year):
-    data=calculate_monthly_costs_per_project(year)
-    bill_data = {}
-    for project in data:
-        if data[project]['total_cost'] > 0:
-            bill_data[project] = data[project]
-    with open('costs.json', 'w') as fp:
-        json.dump(bill_data, fp)
-
-
-yearly_billing_report(2023)
-# TODO: create excel document with all the data
+    if year == today.year:
+        filename = f'output/yearly_cost_report_{year}{today.strftime("%U")}.xlsx'
+    else:
+        filename = f'output/yearly_cost_report_{year}.xlsx'
+    if not os.path.isfile(filename):
+        data=calculate_monthly_costs_per_project(year)
+        bill_data = {}
+        columns = ['project', 'usage_details', 'created', 'deleted', 'owner_name', 'owner_vunetid', 'budget_code', 'budget_type', 'budget_holder', 'total_research_cost', 'total_dataset_cost', 'total_cost', 'yearly_research_cost']
+        
+        for project in data:
+            if data[project]['total_cost'] > 0:
+                bill_data[project] = data[project]
+        with open(f'{filename[:-5]}.json', 'w') as fp:
+            json.dump(bill_data, fp)
+        df = pd.DataFrame.from_dict(data=bill_data, orient='index', columns=columns)
+        df.to_excel(filename)
+    return filename
